refactor(InputCarValidator) : 1주차 피드백 반영, Validator 클래스 따로 적용
